Transcoder: get duration as float in process method
getOutputDuration and getStreamDuration methods return duration as
float.

+++ src/AvTranscoder/transcoder/Transcoder.cpp
@@ -241,14 +241,14 @@ ProcessStat Transcoder::process( IProgress& progress )
 
 	preProcessCodecLatency();
 
-	const double outputDuration = getOutputDuration();
+	const float outputDuration = getOutputDuration();
 	LOG_INFO( "Output duration of the process will be " << outputDuration << "s." )
 
 	size_t frame = 0;
 	bool frameProcessed = true;
 	while( frameProcessed )
 	{
-		const double progressDuration = _outputFile.getStream( 0 ).getStreamDuration();
+		const float progressDuration = _outputFile.getStream( 0 ).getStreamDuration();
 
 		// check if JobStatusCancel
 		if( progress.progress( ( progressDuration > outputDuration ) ? outputDuration : progressDuration, outputDuration ) == eJobStatusCancel )
